Savable hero upgrade
First hero upgraed "Fluffers" is now savable and loadable through the save system

+++ New Unity Project/Assets/Scripts/GameData/UpgradeData.cs
@@ -9,7 +9,7 @@ public class UpgradeData
     public string upgradeName;
     public float currentLevel;
     public float currentDamage;
-    public float heroUnlockOrder;
+    public int heroUnlockOrder;
     public float baseCost;
     public float upgradeIndex;
